{{cookiecutter.app_slug}}/requirements.txt
alembic==1.0.7
bcrypt==3.1.6
blinker==1.4
cffi==1.11.5
Click==7.0
coverage==4.5.1
dominate==2.3.5
flake8==3.5.0
Flask==1.0.2
Flask-Bcrypt==0.7.1
Flask-Bootstrap==3.3.7.1
Flask-DebugToolbar==0.10.1
Flask-Login==0.4.1
Flask-Migrate==2.3.1
Flask-SQLAlchemy==2.3.2
Flask-Testing==0.7.1
Flask-WTF==0.14.2
future==0.17.1
gitdb2==2.0.5
GitPython==2.1.11
itsdangerous==1.1.0
Jinja2==2.10
lightgbm==2.2.3
Mako==1.0.7
MarkupSafe==1.1.0
mccabe==0.6.1
numpy==1.16.1
pandas==0.24.1
psycopg2==2.7.7
pycodestyle==2.3.1
pycparser==2.19
pyflakes==1.6.0
python-dateutil==2.7.5
python-editor==1.0.4
pytz==2018.9
PyYAML==3.13
scikit-learn==0.20.2
scipy==1.2.0
six==1.12.0
smmap2==2.0.5
SQLAlchemy==1.2.17
vertica-python==0.8.2
visitor==0.1.3
Werkzeug==0.14.1
WTForms==2.2.1
xlrd==1.2.0
